# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "For the administration of reports.",
        "title": "Reporting-Service API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "{version}"
    },
    "basePath": "/",
    "paths": {
        "/report": {
            "get": {
                "description": "Gets all reports",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get all reports",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.Report"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates report model",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Update report model",
                "parameters": [
                    {
                        "description": "Report",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Report"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates report model",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Create report model",
                "parameters": [
                    {
                        "description": "Report",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Report"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/:id": {
            "get": {
                "description": "Gets report by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get report by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Report"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes report by id",
                "tags": [
                    "Report"
                ],
                "summary": "Delete report by id",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/create": {
            "post": {
                "description": "Queues creation of a report file. Returns the id of the report and the id of the job that tracks the creation, which can be polled at /report/job/:jobId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Queue report file creation",
                "parameters": [
                    {
                        "description": "Report",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Report"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/file/:reportId/:fileId": {
            "get": {
                "description": "Gets report file by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get report file by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "reportId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes report file by id",
                "tags": [
                    "Report"
                ],
                "summary": "Delete report file by id",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/job": {
            "get": {
                "description": "Gets the most recent report jobs of the user, newest first",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get report jobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "only jobs of this report",
                        "name": "reportId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "maximum number of jobs to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.ReportJob"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report/job/:jobId": {
            "get": {
                "description": "Gets the status of a queued report file creation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get report job by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.ReportJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/templates": {
            "get": {
                "description": "Gets all templates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "Get all templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/lib.Template"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/templates/:id": {
            "get": {
                "description": "Gets template by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "Get template by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Template"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/templates/preview/:id": {
            "get": {
                "description": "Gets template preview by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "Get template preview by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "lib.Data": {
            "type": "object",
            "properties": {
                "dataJsonString": {
                    "type": "string"
                },
                "dataStructured": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/lib.DataType"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "lib.DeviceQuery": {
            "type": "object",
            "properties": {
                "last": {
                    "type": "string"
                }
            }
        },
...
        "lib.Report": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "cron": {
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/lib.ReportObject"
                    }
                },
                "emailHTML": {
                    "type": "string"
                },
                "emailReceivers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "emailSubject": {
                    "type": "string"
                },
                "emailText": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reportFiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.ReportFile"
                    }
                },
                "templateId": {
                    "type": "string"
                },
                "templateName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "lib.ReportFile": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "lib.ReportJob": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reportFileId": {
                    "type": "string"
                },
                "reportId": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/lib.ReportJobStatus"
                },
                "step": {
                    "type": "string"
                }
            }
        },
        "lib.ReportJobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "done",
                "failed"
            ],
            "x-enum-varnames": [
                "ReportJobPending",
                "ReportJobRunning",
                "ReportJobDone",
                "ReportJobFailed"
            ]
        },
        "lib.ReportObject": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/lib.ReportObject"
                    }
                },
                "deviceQuery": {
                    "$ref": "#/definitions/lib.DeviceQuery"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/lib.ReportObject"
                    }
                },
                "length": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "query": {
                    "$ref": "#/definitions/model.QueriesRequestElement"
                },
                "queryOptions": {
                    "$ref": "#/definitions/lib.QueryOptions"
                },
                "value": {},
                "valueType": {
                    "type": "string"
                }
            }
        },
        "lib.Template": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/lib.Data"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.QueriesRequestElement": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QueriesRequestElementColumn"
                    }
                },
                "deviceGroupId": {
                    "type": "string"
                },
                "deviceId": {
                    "type": "string"
                },
                "exportId": {
                    "type": "string"
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QueriesRequestElementFilter"
                    }
                },
                "groupTime": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "locationId": {
                    "type": "string"
                },
                "orderColumnIndex": {
                    "type": "integer"
                },
                "orderDirection": {
                    "$ref": "#/definitions/model.Direction"
                },
                "serviceId": {
                    "type": "string"
                },
                "time": {
                    "$ref": "#/definitions/model.QueriesRequestElementTime"
                }
            }
        },
...
    }
}